updating some tests and organizing archives
