+++ iOS/Sources/Beagle/Sources/Components/Page+Renderable/Tests/PageViewTests.swift
@@ -26,8 +26,6 @@ class PageViewTests: XCTestCase {
             try componentFromJsonFile(componentType: PageView.self, fileName: "PageViewInvalid")
         )
     }
-    
-    private let indicator = PageIndicator(selectedColor: "#d1cebd", unselectedColor: "#f6eedf")
 
     private let page = Container(children: [
         Text("First text"),
@@ -37,18 +35,7 @@ class PageViewTests: XCTestCase {
 
     func test_viewWithPages() {
         let pageView = PageView(
-            children: Array(repeating: page, count: 5),
-            pageIndicator: nil
-        )
-
-        let screen = Beagle.screen(.declarative(pageView.toScreen()))
-        assertSnapshotImage(screen)
-    }
-
-    func test_viewWithPagesAndIndicator() {
-        let pageView = PageView(
-            children: Array(repeating: page, count: 5),
-            pageIndicator: indicator
+            children: Array(repeating: page, count: 5)
         )
 
         let screen = Beagle.screen(.declarative(pageView.toScreen()))
@@ -57,8 +44,7 @@ class PageViewTests: XCTestCase {
     
     func test_viewWithNoPages() {
         let pageView = PageView(
-            children: [],
-            pageIndicator: indicator
+            children: []
         )
 
         let screen = Beagle.screen(.declarative(pageView.toScreen()))
